server: BelaScope: move SC_BelaScope.h to /include
